Move `gatewayclient` code into pkg/indexgateway/
