+++ pkg/indexgateway/index_gateway_grpc_pool.go
@@ -1,4 +1,4 @@
-package gatewayclient
+package indexgateway
 
 import (
 	"io"
